fix: some icon bug
