+++ packages/api/src/commands/index.ts
@@ -1,7 +1,6 @@
 export * from "./apps"
 export * from "./fs"
 export * from "./server"
-export * from "./system"
 export * from "./tools"
 export * from "./extension"
 export * from "./store"
@@ -12,3 +11,4 @@ export * from "./fileSearch"
 export * from "./utils"
 export * as macSecurity from "./mac-security"
 export * from "./mdns"
+export * from "./common"
