+++ packages/api/src/models/icon.ts
@@ -29,8 +29,8 @@ export const BaseIcon = v.object({
 	value: v.string(),
 	invert: v.optional(v.boolean()),
 	darkInvert: v.optional(v.boolean()),
-	hexColor: v.string(),
-	bgColor: v.string()
+	hexColor: v.optional(v.string()),
+	bgColor: v.optional(v.string())
 	// hexColor: v.optional(v.pipe(v.string(), v.hexColor("The hex color is badly formatted."))),
 	// bgColor: v.optional(v.pipe(v.string(), v.hexColor("The hex color is badly formatted.")))
 })
