+++ packages/ui/src/components/main/SystemCmds.svelte
@@ -2,6 +2,7 @@
 	import { CmdTypeEnum, IconEnum, SysCommand } from "@kksh/api/models"
 	import { Command } from "@kksh/svelte5"
 	import { IconMultiplexer } from "@kksh/ui"
+	import { confirm } from "@tauri-apps/plugin-dialog"
 	import { DraggableCommandGroup } from "../custom"
 	import { CmdValue } from "./types"
 
@@ -12,8 +13,13 @@
 	{#each systemCommands as cmd}
 		<Command.Item
 			class="flex justify-between"
-			onSelect={() => {
-				cmd.function()
+			onSelect={async () => {
+				if (cmd.confirmRequired) {
+					const confirmed = await confirm(`Are you sure you want to run ${cmd.name}?`)
+					if (confirmed) {
+						cmd.function()
+					}
+				}
 			}}
 			value={JSON.stringify({
 				cmdName: cmd.name,
